+++ src/com/magento/idea/magento2plugin/magento/packages/MagentoPackages.java
@@ -6,6 +6,11 @@ package com.magento.idea.magento2plugin.magento.packages;
 
 public class MagentoPackages {
     public static String PACKAGES_ROOT = "app/code";
+    public static String AREA_BASE = "base";
     public static String AREA_ADMINHTML = "adminhtml";
     public static String AREA_FRONTEND = "frontend";
+    public static String AREA_CRON = "crontab";
+    public static String AREA_API_REST = "webapi_rest";
+    public static String AREA_API_SOAP = "webapi_soap";
+    public static String AREA_GRAPHQL = "graphql";
 }